+++ src/DPMatrix_GPU.cu
@@ -21,7 +21,7 @@ __device__ __host__ int dp_matrix_calc_delta(int i, int j, int k, int l, sequenc
     return delta_i + delta_k + delta_mi;
 }
 
-__device__ __host__ int dp_matrix_get_pos(int *dp_matrix, const int &i, const int &j, const int &k, const int &l, sequences* seq_ctx)
+__device__ __host__ float dp_matrix_get_pos(float *dp_matrix, const int &i, const int &j, const int &k, const int &l, sequences* seq_ctx)
 {
     if (dp_matrix_check_border(i, j, k, l, seq_ctx) == false)
         return -1024;
@@ -29,16 +29,16 @@ __device__ __host__ int dp_matrix_get_pos(int *dp_matrix, const int &i, const in
     return dp_matrix[dp_matrix_calc_delta(i, j, k, l, seq_ctx)];
 }
 
-__device__ __host__ void dp_matrix_put_pos(int *dp_matrix, const int &i, const int &j, const int &k, const int &l, const int &val, sequences* seq_ctx)
+__device__ __host__ void dp_matrix_put_pos(float *dp_matrix, const int &i, const int &j, const int &k, const int &l, const float &val, sequences* seq_ctx)
 {
     dp_matrix[dp_matrix_calc_delta(i, j, k, l, seq_ctx)] = val;
 }
 
-__host__ int dp_matrix_get_val(int *dp_matrix, const int &i, const int &j, const int &k, const int &l, sequences* seq_ctx)
+__host__ float dp_matrix_get_val(float *dp_matrix, const int &i, const int &j, const int &k, const int &l, sequences* seq_ctx)
 {
     if (dp_matrix_check_border(i, j, k, l, seq_ctx) == false)
         return -1024;
-    int val = 0;
-    cudaMemcpy((void*)(&val), (void*)(&dp_matrix[dp_matrix_calc_delta(i, j, k, l, seq_ctx)]), sizeof(int), cudaMemcpyDeviceToHost);
+    float val = 0;
+    cudaMemcpy((void*)(&val), (void*)(&dp_matrix[dp_matrix_calc_delta(i, j, k, l, seq_ctx)]), sizeof(float), cudaMemcpyDeviceToHost);
     return val;
 }
